requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-ckeditor==6.4.0
django-js-asset==2.0.0
django-widget-tweaks==1.4.12
gunicorn==20.1.0
idna==3.3
Pillow==9.1.0
python-decouple==3.6
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
